Refactor certificate generation logic and add unit tests for self-signed certificate creation
