+++ internal/certs/certify_test.go
@@ -1,4 +1,4 @@
-package proxy
+package certs
 
 import (
 	"os"
